created feedforward-controlled struct RNN
This did extremely well on all four benchmarks. On Match Multi, it took 3 epochs on the first trial, and 4 on the second trial. It got through Repeat in 5 epochs on the first trial. These are very good results, especially for benchmarks which seem to deeply trouble the LSTM. However, it is pretty clear *why* the network does so well: it's only choice is to use structures which were specially crafted for the tasks at hand.

I am eager to test the "ffstruct" model on harder tasks like language modeling. I suspect it will not do so well, since it doesn't have general-purpose human-like memory.